trainer validation step update

+++ Pytorch_Exercise/src_to_implement/trainer.py
@@ -85,7 +85,7 @@ class Trainer:
         # TODO
 
         y_val_pred = self._model(x)
-        loss_val = self._crit(y_val_pred, y)
+        loss_val = self._crit(y_val_pred, y.float())
 
         return loss_val, y_val_pred
